% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate.R
\name{best_model.logitsgl}
\alias{best_model.logitsgl}
\title{Index of best model}
\usage{
\method{best_model}{logitsgl}(object, ...)
}
\arguments{
\item{object}{a logitsgl object}

\item{...}{additional parameters (ignored)}
}
\value{
index of the best model.
}
\description{
Returns the index of the best model, in terms of lowest error rate
}
\examples{
data(birds)

# Simple standardization
X <- scale(X)

lambda <- logitsgl.lambda(X, Y, alpha = 0.5, lambda.min = 0.5)
fit.cv <- logitsgl.cv(X, Y, alpha = 0.5, fold = 5, lambda = lambda)

# print some info
fit.cv

best_model(fit.cv)
}
\author{
Martin Vincent
}

