% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coef.logitsgl}
\alias{coef.logitsgl}
\title{Extract nonzero coefficients}
\usage{
\method{coef}{logitsgl}(object, index = 1:nmod(object), ...)
}
\arguments{
\item{object}{a logitsgl object}

\item{index}{indices of the models}

\item{...}{ignored}
}
\value{
a list of length \code{length(index)} with nonzero coefficients of the models
}
\description{
Extract nonzero coefficients
}
\author{
Martin Vincent
}

