% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitsgl_navigate.R
\name{Err.logitsgl}
\alias{Err.logitsgl}
\title{Compute error rates}
\usage{
\method{Err}{logitsgl}(object, data = NULL, response = object$Y.true,
  Y = response, type = "rate", ...)
}
\arguments{
\item{object}{a logitsgl object}

\item{data}{a design matrix (the \eqn{X} matrix)}

\item{response}{redirected to \code{y}}

\item{Y}{a matrix of the true responses (the \eqn{Y} matrix)}

\item{type}{type of error rate}

\item{...}{ignored}
}
\value{
a vector of error rates
}
\description{
Compute error rates for each model.
If \code{type = "rate"} then the misclassification rates will be computed.
If \code{type = "count"} then the misclassification counts will be computed.
If \code{type = "loglike"} then the negative log likelihood error will be computed.
}
\author{
Martin Vincent
}

