% Generated by roxygen2 (4.0.1): do not edit by hand
\name{features.logitsgl}
\alias{features.logitsgl}
\title{Nonzero features}
\usage{
\method{features}{logitsgl}(object, ...)
}
\arguments{
\item{object}{a logitsgl object}

\item{...}{ignored}
}
\value{
a list of of length \code{nmod(x)} containing the nonzero features (that is nonzero columns of the beta matrices)
}
\description{
Extracts the nonzero features for each model.
}
\author{
Martin Vincent
}

