% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate.R
\name{features_stat.logitsgl}
\alias{features_stat.logitsgl}
\title{Extract feature statistics}
\usage{
\method{features_stat}{logitsgl}(object, ...)
}
\arguments{
\item{object}{a logitsgl object}

\item{...}{ignored}
}
\value{
a vector of length \code{nmod(x)} or a matrix containing the number of nonzero features (or group) of the models.
}
\description{
Extracts the number of nonzero features (or group) in each model.
}
\examples{
data(birds)

# Simple standardization
X <- scale(X)

lambda <- logitsgl.lambda(X, Y, alpha = 0.5, lambda.min = 0.5)
fit <- logitsgl(X, Y, alpha = 0.5, lambda = lambda)
##FIXME
}
\author{
Martin Vincent
}

