% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitsgl_config.R
\name{logitsgl.c.config}
\alias{logitsgl.c.config}
\title{Featch information about the C side configuration of the package}
\usage{
logitsgl.c.config()
}
\value{
list
}
\description{
Featch information about the C side configuration of the package
}
\author{
Martin Vicnet
}

