% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{logitsgl}
\alias{logitsgl}
\title{Multi-label Sparse Group Lasso Logistic Regression}
\usage{
logitsgl(x, y, intercept = TRUE, grouping = factor(1:ncol(x)),
  groupWeights = NULL, parameterWeights = NULL, alpha = 1, lambda,
  algorithm.config = logitsgl.standard.config)
}
\arguments{
\item{x}{design matrix, matrix of size \eqn{N \times p}.}

\item{y}{response matrix, matrix of size \eqn{N \times K}.}

\item{intercept}{should the model include intercept parameters}

\item{grouping}{grouping of features, a factor or vector of length \eqn{p}.
Each element of the factor/vector specifying the group of the feature.}

\item{groupWeights}{the group weights, a vector of length \eqn{m} (the number of groups).}

\item{parameterWeights}{a matrix of size \eqn{K \times p}.}

\item{alpha}{the \eqn{\alpha} value 0 for group lasso, 1 for lasso, between 0 and 1 gives a sparse group lasso penalty.}

\item{lambda}{lambda sequence.}

\item{algorithm.config}{the algorithm configuration to be used.}
}
\value{
\item{beta}{the fitted parameters -- the list \eqn{\hat\beta(\lambda(1)), \dots, \hat\beta(\lambda(d))} of length \code{length(return)}.
With each entry of list holding the fitted parameters, that is matrices of size \eqn{K\times p} (if \code{intercept = TRUE} matrices of size \eqn{K\times (p+1)})}
\item{loss}{the values of the loss function.}
\item{objective}{the values of the objective function (i.e. loss + penalty).}
\item{lambda}{the lambda values used.}
}
\description{
Fit a multi-label sparse group lasso logistic regression using sparse group lasso.
}
\examples{
data(birds)

# Simple standardization
X <- scale(X)

lambda <- logitsgl.lambda(X, Y, alpha = 0.5, lambda.min = 0.5)

fit <- logitsgl(X, Y, alpha = 0.5, lambda = lambda)

# print some information
fit

# Traning error
Err(fit, X)
}
\author{
Martin Vincent
}

