% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{logitsgl.standard.config}
\alias{logitsgl.standard.config}
\title{Standard algorithm configuration}
\format{An object of class \code{list} of length 15.}
\usage{
logitsgl.standard.config
}
\description{
\code{logitsgl.standard.config <- logitsgl.algorithm.config()}
}
\author{
Martin Vicnet
}
\keyword{datasets}

