% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling.R
\name{logitsgl.subsampling}
\alias{logitsgl.subsampling}
\title{Subsampling}
\usage{
logitsgl.subsampling(x, y, intercept = TRUE, grouping = factor(1:ncol(x)),
  groupWeights = c(sqrt(ncol(y) * table(grouping))),
  parameterWeights = matrix(1, nrow = ncol(y), ncol = ncol(x)), alpha = 1,
  lambda, train, test, collapse = FALSE, use_parallel = FALSE,
  algorithm.config = logitsgl.standard.config)
}
\arguments{
\item{x}{design matrix, matrix of size \eqn{N \times p}.}

\item{y}{response matrix, matrix of size \eqn{N \times K}.}

\item{intercept}{should the model include intercept parameters}

\item{grouping}{grouping of features, a factor or vector of length \eqn{p}.
Each element of the factor/vector specifying the group of the feature.}

\item{groupWeights}{the group weights, a vector of length \eqn{m} (the number of groups).}

\item{parameterWeights}{a matrix of size \eqn{K \times p}.}

\item{alpha}{the \eqn{\alpha} value 0 for group lasso, 1 for lasso,
between 0 and 1 gives a sparse group lasso penalty.}

\item{lambda}{the lambda sequence for the regularization path.}

\item{train}{a list of training samples,
each item of the list corresponding to a subsample.
Each item in the list must be a vector with the indices of the training samples
for the corresponding subsample.
The length of the list must equal the length of the \code{test} list.}

\item{test}{a list of test samples, each item of the list corresponding to a subsample.
Each item in the list must be vector with the indices of the test samples for the corresponding subsample.
The length of the list must equal the length of the \code{training} list.}

\item{collapse}{if \code{TRUE} the results for each subsample will be collapse into one result (this is useful if the subsamples are not overlapping)}

\item{use_parallel}{If \code{TRUE} the \code{foreach} loop will use \code{\%dopar\%}. The user must registre the parallel backend.}

\item{algorithm.config}{the algorithm configuration to be used.}
}
\value{
\item{Yhat}{TODO}
\item{Y.true}{TODO}
\item{features}{number of features used in the models}
\item{parameters}{number of parameters used in the models.}
}
\description{
Subsampling using multiple possessors
}
\examples{
data(birds)

# Simple standardization
X <- scale(X)

train <- list(1:250, 1:300)
test <- list(301:322, 301:322)

lambda <- lapply(train, function(idx)
	logitsgl.lambda(X[idx,], Y[idx,], alpha = 0.5, lambda.min = 0.4))

# registre parallel backend
cl <- makeCluster(2)
registerDoParallel(cl)

# Do cross validation using 2 threads
fit.sub <- logitsgl.subsampling(X, Y,
	alpha = 0.5,
	train = train,
	test = test,
	lambda = lambda,
	use_parallel = TRUE)

stopCluster(cl)

# print some information
fit.sub
}
\author{
Martin Vincent
}

