% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{logitsgl.test}
\alias{logitsgl.test}
\title{Test objectives}
\usage{
logitsgl.test(x, y, grouping = factor(1:ncol(x)), groupWeights = c(sqrt(2 *
  ncol(y) * table(grouping))), parameterWeights = matrix(1, nrow = (ncol(y)^2
  - ncol(y))/2 + 2 * ncol(y), ncol = ncol(x)))
}
\arguments{
\item{x}{design matrix, matrix of size \eqn{N \times p}.}

\item{y}{response matrix, matrix of size \eqn{N \times K}.}

\item{grouping}{grouping of features, a factor or vector of length \eqn{p}. Each element of the factor/vector specifying the group of the feature.}

\item{groupWeights}{the group weights, a vector of length \eqn{m} (the number of groups).}

\item{parameterWeights}{a matrix of size \eqn{K \times p}.}
}
\value{
TODO
}
\description{
TODO description
}
\author{
Martin Vincent
}

