% Generated by roxygen2 (4.0.1): do not edit by hand
\name{models.logitsgl}
\alias{models.logitsgl}
\title{Exstract the fitted models}
\usage{
\method{models}{logitsgl}(object, index = 1:nmod(object), ...)
}
\arguments{
\item{object}{a logitsgl object}

\item{index}{indices of the models to be returned}

\item{...}{ignored}
}
\value{
a list of \eqn{\beta} matrices.
}
\description{
Returns the fitted models, that is the estimated \eqn{\beta} matrices.
}
\author{
Martin Vincent
}

