% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parameters.logitsgl}
\alias{parameters.logitsgl}
\title{Nonzero parameters}
\usage{
\method{parameters}{logitsgl}(object, ...)
}
\arguments{
\item{object}{a logitsgl object}

\item{...}{ignored}
}
\value{
a list of length \code{nmod(x)} containing the nonzero parameters of the models.
}
\description{
Extracts the nonzero parameters for each model.
}
\author{
Martin Vincent
}

