% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate.R
\name{parameters_stat.logitsgl}
\alias{parameters_stat.logitsgl}
\title{Extracting parameter statistics}
\usage{
\method{parameters_stat}{logitsgl}(object, ...)
}
\arguments{
\item{object}{a logitsgl object}

\item{...}{ignored}
}
\value{
a vector of length \code{nmod(x)} or a matrix containing the number of nonzero parameters of the models.
}
\description{
Extracts the number of nonzero parameters in each model.
}
\examples{
data(birds)

# Simple standardization
X <- scale(X)

lambda <- logitsgl.lambda(X, Y, alpha = 0.5, lambda.min = 0.5)
fit <- logitsgl(X, Y, alpha = 0.5, lambda = lambda)

# Number of parameters in models
parameters_stat(fit)

fit.cv <- logitsgl.cv(X, Y, alpha = 0.5, fold = 2, lambda = lambda)

# Number of parameters
parameters_stat(fit.cv)
}
\author{
Martin Vincent
}

