% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.logitsgl}
\alias{predict.logitsgl}
\title{Predict}
\usage{
\method{predict}{logitsgl}(object, x, sparse.data = is(x, "sparseMatrix"),
  ...)
}
\arguments{
\item{object}{an object of class logitsgl, produced with \code{logitsgl}.}

\item{x}{a data matrix of size \eqn{N_\textrm{new} \times p}.}

\item{sparse.data}{if TRUE \code{x} will be treated as sparse, if \code{x} is a sparse matrix it will be treated as sparse by default.}

\item{...}{ignored.}
}
\value{
\item{Yhat}{the predicted response matrix (of size \eqn{N_\textrm{new} \times K})}
}
\description{
Compute the predicted response matrix for a new data set.
}
\author{
Martin Vincent
}

