% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_objective_logitsgl}
\alias{test_objective_logitsgl}
\title{Test objectives}
\usage{
test_objective_logitsgl(x, y, grouping = factor(1:ncol(x)),
  groupWeights = c(sqrt(ncol(y) * table(grouping))),
  parameterWeights = matrix(1, nrow = ncol(y), ncol = ncol(x)))
}
\arguments{
\item{x}{design matrix, matrix of size \eqn{N \times p}.}

\item{y}{response matrix, matrix of size \eqn{N \times K}.}

\item{grouping}{grouping of features, a factor or vector of length \eqn{p}. Each element of the factor/vector specifying the group of the feature.}

\item{groupWeights}{the group weights, a vector of length \eqn{m} (the number of groups).}

\item{parameterWeights}{a matrix of size \eqn{K \times p}.}
}
\value{
number of detected problems
}
\description{
TODO description
}
\examples{
data(birds)

# scale down
X <- X[1:100, 1:10]
Y <- Y[1:100, 1:3]

# Simple standardization
X <- scale(X)

# look for problems in objective -- this may be slow for high dimensions
n_problems_detected <- test_objective_logitsgl(X, Y)

if(n_problems_detected > 0) stop("problem detected in objective")
}
\author{
Martin Vincent
}

