% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitsgl_navigate.R
\name{coef.logitsgl}
\alias{coef.logitsgl}
\title{Extract nonzero coefficients}
\usage{
\method{coef}{logitsgl}(object, index = 1:nmod(object), ...)
}
\arguments{
\item{object}{a logitsgl object}

\item{index}{indices of the models}

\item{...}{ignored}
}
\value{
a list of length \code{length(index)} with nonzero coefficients of the models
}
\author{
Martin Vincent
}

