% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Err.logitsgl}
\alias{Err.logitsgl}
\title{Compute error rates}
\usage{
\method{Err}{logitsgl}(object, data = NULL, response = object$Y.true,
  Y = response, type = "rate", ...)
}
\arguments{
\item{object}{a logitsgl object}

\item{data}{a design matrix (the \eqn{X} matrix)}

\item{response}{redirected to \code{y}}

\item{y}{a matrix of the true responses (the \eqn{Y} matrix)}

\item{...}{ignored}
}
\value{
a vector of error rates
}
\description{
Compute and return the error
TODO
for each model.
}
\author{
Martin Vincent
}

