% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate.R
\name{features.logitsgl}
\alias{features.logitsgl}
\title{Nonzero features}
\usage{
\method{features}{logitsgl}(object, ...)
}
\arguments{
\item{object}{a logitsgl object}

\item{...}{ignored}
}
\value{
a list of of length \code{nmod(x)} containing the nonzero features (that is nonzero columns of the beta matrices)
}
\description{
Extracts the nonzero features for each model.
}
\examples{
data(birds)

# Simple standardization
X <- scale(X)

lambda <- logitsgl.lambda(X, Y, alpha = 0.5, lambda.min = 0.5)
fit <- logitsgl(X, Y, alpha = 0.5, lambda = lambda)

# the nonzero features of model 1, 10 and 25
features(fit)[c(1,10,25)]

# count the number of nonzero features in each model
sapply(features(fit), length)
}
\author{
Martin Vincent
}

