% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate.R
\name{models.logitsgl}
\alias{models.logitsgl}
\title{Exstract the fitted models}
\usage{
\method{models}{logitsgl}(object, index = 1:nmod(object), ...)
}
\arguments{
\item{object}{a logitsgl object}

\item{index}{indices of the models to be returned}

\item{...}{ignored}
}
\value{
a list of \eqn{\beta} matrices.
}
\description{
Returns the fitted models, that is the estimated \eqn{\beta} matrices.
}
\examples{
data(birds)

# Simple standardization
X <- scale(X)

lambda <- logitsgl.lambda(X, Y, alpha = 0.5, lambda.min = 0.5)
fit <- logitsgl(X, Y, alpha = 0.5, lambda = lambda)

# Exstract model 50
models(fit)[[50]]
}
\author{
Martin Vincent
}

