% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate.R
\name{nmod.logitsgl}
\alias{nmod.logitsgl}
\title{Number of models used for fitting}
\usage{
\method{nmod}{logitsgl}(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{additional parameters (optional)}
}
\value{
the number of models used when fitting the object \code{x}.
}
\description{
Returns the number of models used for fitting.
However, note that the objects returned by \code{msgl.cv} and
\code{msgl.subsampling} does not contain any models even though
\code{nmod} returns a nonzero number.
}
\author{
Martin Vincent
}

