% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logitsgl_navigate.R
\name{nmod.logitsgl}
\alias{nmod.logitsgl}
\title{Returns the number of models in a lsgl object}
\usage{
\method{nmod}{logitsgl}(object, ...)
}
\arguments{
\item{object}{a logitsgl object}

\item{...}{ignored}
}
\value{
the number of models in \code{object}
}
\description{
Returns the number of models in a lsgl object
}
\author{
Martin Vincent
}

