% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate.R
\name{parameters.logitsgl}
\alias{parameters.logitsgl}
\title{Nonzero parameters}
\usage{
\method{parameters}{logitsgl}(object, ...)
}
\arguments{
\item{object}{a logitsgl object}

\item{...}{ignored}
}
\value{
a list of length \code{nmod(x)} containing the nonzero parameters of the models.
}
\description{
Extracts the nonzero parameters for each model.
}
\examples{
data(birds)

# Simple standardization
X <- scale(X)

lambda <- logitsgl.lambda(X, Y, alpha = 0.5, lambda.min = 0.6)
fit <- logitsgl(X, Y, alpha = 0.5, lambda = lambda)

# Nonzero parameters for model 100
parameters(fit)[[100]]
}
\author{
Martin Vincent
}

