% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitsgl_predict.R
\name{predict.logitsgl}
\alias{predict.logitsgl}
\title{Predict}
\usage{
\method{predict}{logitsgl}(object, x, sparse.data = is(x, "sparseMatrix"),
  ...)
}
\arguments{
\item{object}{an object of class logitsgl, produced with \code{logitsgl}.}

\item{x}{a data matrix of size \eqn{N_\textrm{new} \times p}.}

\item{sparse.data}{if TRUE \code{x} will be treated as sparse, if \code{x} is a sparse matrix it will be treated as sparse by default.}

\item{...}{ignored.}
}
\value{
\item{Yhat}{the predicted response matrix (of size \eqn{N_\textrm{new} \times K})}
}
\description{
Compute the predicted response matrix for a new data set.
}
\author{
Martin Vincent
data(birds)

# Simple standardization
X <- scale(X)

lambda <- logitsgl.lambda(X, Y, alpha = 0.5, lambda.min = 0.5)

fit <- logitsgl(X, Y, alpha = 0.5, lambda = lambda)

# training errors
res <- predict(fit, X)
}

