% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logitsgl_navigate.R
\name{print.logitsgl}
\alias{print.logitsgl}
\title{Print function for logitsgl}
\usage{
\method{print}{logitsgl}(x, ...)
}
\arguments{
\item{x}{logitsgl object}

\item{...}{ignored}
}
\description{
This function will print some general information about the lsgl object
}
\author{
Martin Vincent
}

